OO-565 Sume of Sammelbestellung; Recalculate Bestellung on change in Status Abgeschlossen
