+++ src/main/scala/ch/openolitor/core/db/evolution/scripts/Scripts.scala
@@ -54,5 +54,6 @@ object Scripts {
       OO506_DBScripts.scripts ++
       OO501_DBScripts.scripts ++
       OO544_DBScripts.scripts ++
-      OO564_DBScripts.scripts
+      OO564_DBScripts.scripts ++
+      OO544_DBScripts.scripts
 }
