+++ src/main/scala/ch/openolitor/stammdaten/StammdatenInsertService.scala
@@ -61,6 +61,7 @@ class StammdatenInsertService(override val sysConfig: SystemConfig) extends Even
     with AsyncConnectionPoolContextAware
     with StammdatenDBMappings
     with KorbHandler
+    with SammelbestellungenHandler
     with LieferungHandler {
   self: StammdatenWriteRepositoryComponent =>
 
@@ -662,100 +663,7 @@ class StammdatenInsertService(override val sysConfig: SystemConfig) extends Even
 
   def createSammelbestellungen(meta: EventMetadata, id: SammelbestellungId, create: SammelbestellungCreate)(implicit personId: PersonId = meta.originator) = {
     DB autoCommit { implicit session =>
-      stammdatenWriteRepository.getById(produzentMapping, create.produzentId) map { produzent =>
-        val sammelbestellung = Sammelbestellung(
-          id,
-          create.produzentId,
-          produzent.kurzzeichen,
-          create.lieferplanungId,
-          Abgeschlossen,
-          create.datum,
-          None,
-          None,
-          0,
-          produzent.mwstSatz,
-          0,
-          0,
-          DateTime.now,
-          personId,
-          DateTime.now,
-          personId
-        )
-
-        stammdatenWriteRepository.insertEntity[Sammelbestellung, SammelbestellungId](sammelbestellung) map { sammelbestellung =>
-
-          // delete all Bestellpositionen from Bestellungen (Bestellungen are maintained even if nothing is ordered/bestellt)
-          stammdatenWriteRepository.getBestellpositionenBySammelbestellung(id) foreach {
-            position => stammdatenWriteRepository.deleteEntity[Bestellposition, BestellpositionId](position.id)
-          }
-
-          stammdatenWriteRepository.getLieferungenDetails(create.lieferplanungId) groupBy (_.abotyp.get.adminProzente) map {
-            case (adminProzente, lieferungen) =>
-              val anzahlKoerbeZuLiefern = lieferungen.map(l => (l.id, l.anzahlKoerbeZuLiefern)).toMap
-
-              val bestellung = Bestellung(
-                BestellungId(IdUtil.positiveRandomId),
-                id,
-                0,
-                produzent.mwstSatz,
-                0,
-                0,
-                adminProzente,
-                0,
-                0,
-                DateTime.now,
-                personId,
-                DateTime.now,
-                personId
-              )
-
-              stammdatenWriteRepository.insertEntity[Bestellung, BestellungId](bestellung) map { bestellung =>
-                val positionen = stammdatenWriteRepository.getLieferpositionenByLieferplanAndProduzent(create.lieferplanungId, create.produzentId).
-                  //group by same produkt, menge and preis
-                  groupBy(x => (x.produktId, x.menge, x.preis)).map {
-                    case ((produktId, menge, preis), positionen) =>
-                      positionen.map(lp => anzahlKoerbeZuLiefern.get(lp.lieferungId).getOrElse(0)).sum match {
-                        case 0 => //don't add position
-                          None
-                        case anzahl =>
-                          positionen.headOption map { lieferposition =>
-                            Bestellposition(
-                              BestellpositionId(IdUtil.positiveRandomId),
-                              bestellung.id,
-                              lieferposition.produktId,
-                              lieferposition.produktBeschrieb,
-                              lieferposition.preisEinheit,
-                              lieferposition.einheit,
-                              menge.getOrElse(0),
-                              preis.map(_ * anzahl),
-                              anzahl,
-                              DateTime.now,
-                              personId,
-                              DateTime.now,
-                              personId
-                            )
-                          }
-                      }
-                  }.flatten
-
-                positionen.map { bestellposition =>
-                  stammdatenWriteRepository.insertEntity[Bestellposition, BestellpositionId](bestellposition)
-                }
-
-                val total = positionen.map(_.preis).flatten.sum
-                val adminProzenteAbzug = bestellung.adminProzente / 100 * total
-                val totalNachAbzugAdminProzente = total - adminProzenteAbzug
-                // mwst auf total ohne adminanteil
-                val mwst = bestellung.steuerSatz.map(_ / 100 * totalNachAbzugAdminProzente).getOrElse(BigDecimal(0))
-                val totalInkl = totalNachAbzugAdminProzente + mwst
-
-                //update total on bestellung, steuer and totalSteuer
-                val copy = bestellung.copy(preisTotal = total, steuer = mwst, totalSteuer = totalInkl, adminProzenteAbzug = adminProzenteAbzug, totalNachAbzugAdminProzente = totalNachAbzugAdminProzente)
-                stammdatenWriteRepository.updateEntity[Bestellung, BestellungId](copy)
-              }
-          }
-        }
-      }
+      createOrUpdateSammelbestellungen(id, create)
     }
   }
