+++ src/main/scala/ch/openolitor/stammdaten/StammdatenUpdateService.scala
@@ -55,7 +55,10 @@ class DefaultStammdatenUpdateService(sysConfig: SystemConfig, override val syste
 /**
  * Actor zum Verarbeiten der Update Anweisungen innerhalb des Stammdaten Moduls
  */
-class StammdatenUpdateService(override val sysConfig: SystemConfig) extends EventService[EntityUpdatedEvent[_, _]] with LazyLogging with AsyncConnectionPoolContextAware with StammdatenDBMappings {
+class StammdatenUpdateService(override val sysConfig: SystemConfig) extends EventService[EntityUpdatedEvent[_, _]]
+    with LazyLogging
+    with AsyncConnectionPoolContextAware
+    with StammdatenDBMappings {
   self: StammdatenWriteRepositoryComponent =>
 
   val FALSE = false
