+++ src/main/scala/ch/openolitor/stammdaten/repositories/StammdatenWriteRepositories.scala
@@ -83,6 +83,7 @@ trait StammdatenWriteRepository extends BaseWriteRepository with EventStream {
   def getGeplanteLieferungNachher(vertriebId: VertriebId, datum: DateTime)(implicit session: DBSession): Option[Lieferung]
   def countEarlierLieferungOffen(id: LieferplanungId)(implicit session: DBSession): Option[Int]
   def getSammelbestellungen(id: LieferplanungId)(implicit session: DBSession): List[Sammelbestellung]
+  def getSammelbestellungen(id: LieferungId)(implicit session: DBSession): List[Sammelbestellung]
   def getBestellungen(id: SammelbestellungId)(implicit session: DBSession): List[Bestellung]
   def getBestellpositionen(id: BestellungId)(implicit session: DBSession): List[Bestellposition]
   def getBestellpositionenBySammelbestellung(id: SammelbestellungId)(implicit session: DBSession): List[Bestellposition]
@@ -385,6 +386,10 @@ trait StammdatenWriteRepositoryImpl extends StammdatenWriteRepository with LazyL
   }
 
   def getSammelbestellungen(id: LieferplanungId)(implicit session: DBSession): List[Sammelbestellung] = {
+    getSammelbestellungenQuery(id)()
+  }
+
+  def getSammelbestellungen(id: LieferungId)(implicit session: DBSession): List[Sammelbestellung] = {
     getSammelbestellungenQuery(id)()
   }
